APPDEV-3340: Customer order ETA window [Email + Manifest]
git-svn-id: https://appdevsvn.nj01/appdev/FreshDirect/trunk@26737 842219ff-9341-0410-a943-573460bdd4de

+++ projects/TransportationAdmin/src/com/freshdirect/transadmin/web/DomainController.java
@@ -478,7 +478,9 @@ public class DomainController extends AbstractMultiActionController {
 			}
 			_tmpZone.setAmZoneSupervisors(amZoneSupervisors);
 			_tmpZone.setPmZoneSupervisors(pmZoneSupervisors);
-			zoneManagerService.removeEntity(clearSupervisorLst);
+			if(clearSupervisorLst.size() > 0){
+				zoneManagerService.removeEntity(clearSupervisorLst);
+			}
 		} catch (Exception e) {				
 			e.printStackTrace();
 		}
